Updating Jsons
Each collection will have two new fields:
Capitals and Plus.
Plus will refer to whether the country is fully recognised or not. Plus false will have countries proper - those that are recognised and are always included in the continents.
Plus true will have countries which are de-facto independent but not always recognised, countries which are not always included in the continent, as well as some famous or bigger autonomies and territories